+++ hosting/src/App.vue
@@ -1,42 +1,6 @@
 <template>
-  <v-app>
-    <v-app-bar app flat dense>
-      <v-btn name="open menu" icon text @click="drawer = !drawer">
-        <v-img src="./assets/logo.color.svg" height="32" contain />
-      </v-btn>
-      <v-toolbar-title class="pa-0">Score Manager</v-toolbar-title>
-      <v-spacer />
-      <account-button />
-    </v-app-bar>
-
-    <v-main>
-      <v-container fluid class="pb-12 mb-4">
-        <score-table />
-      </v-container>
-    </v-main>
-  </v-app>
+  <router-view />
 </template>
-
-<script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-// import ScoreTable from './components/ScoreTable.vue'
-const ScoreTable = () => import('./components/ScoreTable.vue');
-// import AccountButton from './components/AccountButton.vue'
-const AccountButton = () => import('./components/AccountButton.vue');
-// import LinkList from './components/LinkList.vue'
-const LinkList = () => import('./components/LinkList.vue');
-
-@Component({
-  components: {
-    ScoreTable,
-    AccountButton,
-    LinkList,
-  },
-})
-export default class App extends Vue {
-  drawer = false;
-}
-</script>
 
 <style>
 @import url("https://fonts.googleapis.com/css?family=Noto+Sans+JP|Roboto&display=swap");
