+++ hosting/src/main.ts
@@ -5,10 +5,12 @@ import vuetify from './plugins/vuetify';
 
 import './utils/firebase';
 import './utils/utils';
+import router from './router';
 
 Vue.config.productionTip = false;
 
 new Vue({
   vuetify,
+  router,
   render: (h) => h(App),
 }).$mount('#app');
